Add RouterName for using existing router

+++ pkg/cloud/services/networking/router.go
@@ -45,13 +45,23 @@ func (s *Service) ReconcileRouter(openStackCluster *infrav1.OpenStackCluster, cl
 	}
 
 	routerName := getRouterName(clusterName)
-	s.scope.Logger().Info("Reconciling router", "name", routerName)
+	existingRouter := false
+	if openStackCluster.Spec.RouterName != "" {
+		routerName = openStackCluster.Spec.RouterName
+		existingRouter = true
+	} else {
+		s.scope.Logger().Info("Reconciling router", "name", routerName)
+	}
 
 	routerList, err := s.client.ListRouter(routers.ListOpts{
 		Name: routerName,
 	})
 	if err != nil {
 		return err
+	}
+
+	if existingRouter && len(routerList) == 0 {
+		return fmt.Errorf("router %s not found", routerName)
 	}
 
 	if len(routerList) > 1 {
@@ -190,7 +200,20 @@ func (s *Service) setRouterExternalIPs(openStackCluster *infrav1.OpenStackCluste
 }
 
 func (s *Service) DeleteRouter(openStackCluster *infrav1.OpenStackCluster, clusterName string) error {
-	router, subnet, err := s.getRouter(clusterName)
+	routerName := getRouterName(clusterName)
+	existingRouter := false
+	if openStackCluster.Spec.RouterName != "" {
+		routerName = openStackCluster.Spec.RouterName
+		existingRouter = true
+	}
+
+	router, err := s.getRouterByName(routerName)
+	if err != nil {
+		return err
+	}
+
+	subnetName := getSubnetName(clusterName)
+	subnet, err := s.getSubnetByName(subnetName)
 	if err != nil {
 		return err
 	}
@@ -213,6 +236,11 @@ func (s *Service) DeleteRouter(openStackCluster *infrav1.OpenStackCluster, clust
 		}
 	}
 
+	if existingRouter {
+		s.scope.Logger().V(4).Info("No need to delete pre-existing router", "name", router.Name)
+		return nil
+	}
+
 	err = s.client.DeleteRouter(router.ID)
 	if err != nil {
 		record.Warnf(openStackCluster, "FailedDeleteRouter", "Failed to delete router %s with id %s: %v", router.Name, router.ID, err)
@@ -227,22 +255,6 @@ func (s *Service) getRouterInterfaces(routerID string) ([]ports.Port, error) {
 	return s.client.ListPort(ports.ListOpts{
 		DeviceID: routerID,
 	})
-}
-
-func (s *Service) getRouter(clusterName string) (routers.Router, subnets.Subnet, error) {
-	routerName := getRouterName(clusterName)
-	router, err := s.getRouterByName(routerName)
-	if err != nil {
-		return routers.Router{}, subnets.Subnet{}, err
-	}
-
-	subnetName := getSubnetName(clusterName)
-	subnet, err := s.getSubnetByName(subnetName)
-	if err != nil {
-		return router, subnets.Subnet{}, err
-	}
-
-	return router, subnet, nil
 }
 
 func (s *Service) getRouterByName(routerName string) (routers.Router, error) {
